+++ Tests/LoadFeedFromRemoteUseCaseTests.swift
@@ -5,82 +5,6 @@
 import XCTest
 import FeedAPIChallenge
 
-protocol HTTPClient {
-    typealias Result = Swift.Result<(Data, HTTPURLResponse), Error>
-    
-    func get(from url: URL, completion: @escaping (Result) -> Void)
-}
-
-class RemoteFeedLoader {
-    private let url: URL
-    private let client: HTTPClient
-    
-    enum Error: Swift.Error {
-        case connectivity
-        case invalidData
-    }
-    
-    typealias Result = FeedLoader.Result
-    
-    init(url: URL, client: HTTPClient) {
-        self.url = url
-        self.client = client
-    }
-    
-    func load(completion: @escaping (Result) -> Void) {
-        client.get(from: url) { [weak self] result in
-            guard self != nil else { return }
-            
-            switch result {
-            case .failure:
-                completion(.failure(Error.connectivity))
-            case let .success((data, response)):
-                completion(FeedItemsMapper.map(data, response))
-            }
-        }
-    }
-}
-
-final class FeedItemsMapper {
-    private struct Root: Decodable {
-        let items: [Item]
-        
-        var feed: [FeedImage] {
-            return items.map {
-                FeedImage(id: $0.id,
-                          description: $0.description,
-                          location: $0.location,
-                          url: $0.imageURL
-                )
-            }
-        }
-    }
-    
-    private struct Item: Decodable {
-        let id: UUID
-        let description: String?
-        let location: String?
-        let imageURL: URL
-        
-        enum CodingKeys: String, CodingKey {
-            case id = "image_id"
-            case description = "image_desc"
-            case location = "image_loc"
-            case imageURL = "image_url"
-        }
-    }
-    
-    private static var OK_200: Int { return 200 }
-    
-    static func map(_ data: Data, _ response: HTTPURLResponse) -> RemoteFeedLoader.Result {
-        guard response.statusCode == OK_200,
-            let root = try? JSONDecoder().decode(Root.self, from: data) else {
-                return .failure(RemoteFeedLoader.Error.invalidData)
-        }
-        return .success(root.feed)
-    }
-}
-
 class HTTPClientSpy: HTTPClient {
     typealias Result = (Swift.Result<(Data, HTTPURLResponse), Error>)
     var messages = [(url: URL, completion: (Result) -> Void)]()
